[GJ]part4.21] Add user auth for blog delete route

+++ part4/blog/controllers/blogs.js
@@ -4,14 +4,6 @@ const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
-// const getTokenFrom = (request) => {
-//  const authorization = request.get('authorization');
-//  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-//    return authorization.substring(7);
-//  }
-//  return null;
-// };
-
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 });
   response.json(blogs);
@@ -53,10 +45,17 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+      return response.status(404).end();
+    }
+    if (blog.user.toString() !== request.user.id.toString()) {
+      return response.status(401).end();
+    }
     await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    return response.status(204).end();
   } catch (exception) {
-    next(exception);
+    return next(exception);
   }
 });
